test-vertex-ai/requirements.txt
annotated-types==0.7.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
google-auth==2.37.0
google-genai==0.2.2
idna==3.10
pillow==11.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.3
pydantic_core==2.27.1
requests==2.32.3
rsa==4.9
typing_extensions==4.12.2
urllib3==2.2.3
websockets==14.1
